refactor: spit layout-components into separate component files with classes
