+++ exam.html
@@ -43,7 +43,10 @@
   <script src="static/js/constants.js"></script>
   <script src="static/js/helpers.js"></script>
   <script src="static/js/vfs.js"></script>
-  <script src="static/js/layout-components.js"></script>
+  <script src="static/js/layout/helpers.js"></script>
+  <script src="static/js/layout/term.component.js"></script>
+  <script src="static/js/layout/editor.component.js"></script>
+  <script src="static/js/layout/layout.js"></script>
   <script src="static/js/worker-api.js"></script>
   <script src="static/js/main.shared.js"></script>
   <script src="static/js/main.exam.js"></script>
