+++ index.html
@@ -131,7 +131,11 @@
   <script src="static/js/helpers.js"></script>
   <script src="static/js/modal.js"></script>
   <script src="static/js/vfs.js"></script>
-  <script src="static/js/layout-components.js"></script>
+  <script src="static/js/layout/helpers.js"></script>
+  <script src="static/js/layout/term.component.js"></script>
+  <script src="static/js/layout/editor.component.js"></script>
+  <script src="static/js/layout/layout.js"></script>
+  <script src="static/js/layout/layout.ide.js"></script>
   <script src="static/js/worker-api.js"></script>
   <script src="static/js/gitfs.js"></script>
   <script src="static/js/ide/menubar.js"></script>
